+++ foxy_farmer/foundation/util/bech32_address.py
@@ -1,0 +1,10 @@
+from chia.util.bech32m import decode_puzzle_hash
+
+
+def is_valid_address(address: str) -> bool:
+    try:
+        decode_puzzle_hash(address)
+    except ValueError:
+        return False
+
+    return True