+++ foxy_farmer/pool/pool_joiner.py
@@ -1,17 +1,14 @@
-from asyncio import sleep
 from pathlib import Path
-from typing import Dict, Any, Optional, List
+from typing import Dict, Any, Optional
 
 from chia.cmds.cmds_util import get_wallet
-from chia.rpc.wallet_rpc_client import WalletRpcClient
-from chia.util.config import load_config
-from chia.util.ints import uint16, uint64
-from yaspin import yaspin
+from chia.util.ints import uint64
 
 from foxy_farmer.farmer.embedded_chia_environment import EmbeddedChiaEnvironment
 from foxy_farmer.foundation.wallet.pool_join import get_plot_nft_not_pooling_with_foxy, \
-    await_launcher_pool_join_completion, join_plot_nfts_to_pool
+    await_launcher_pool_join_completion, join_plot_nfts_to_pool, update_foxy_config_plot_nfts_if_required
 from foxy_farmer.foundation.wallet.sync import wait_for_wallet_sync
+from foxy_farmer.foundation.wallet.run_wallet import run_wallet
 from foxy_farmer.foxy_config_manager import FoxyConfigManager
 
 
@@ -27,36 +24,7 @@ class PoolJoiner:
         self._foxy_config_manager = FoxyConfigManager(config_path)
 
     async def join_pool(self, fee: uint64):
-        self._chia_environment = EmbeddedChiaEnvironment(
-            root_path=self._foxy_root,
-            config=self._config,
-            allow_connecting_to_existing_daemon=True,
-        )
-        wallet_rpc: Optional[WalletRpcClient] = None
-
-        try:
-            await self._chia_environment.start_daemon()
-            await self._chia_environment.start_services(["wallet"])
-
-            wallet_rpc = await WalletRpcClient.create(
-                self._config["self_hostname"],
-                uint16(self._config["wallet"]["rpc_port"]),
-                self._foxy_root,
-                self._config,
-            )
-
-            async def is_wallet_reachable() -> bool:
-                try:
-                    await wallet_rpc.healthz()
-
-                    return True
-                except:
-                    return False
-
-            with yaspin(text="Waiting for the wallet to finish starting ..."):
-                while not await is_wallet_reachable():
-                    await sleep(3)
-
+        async with run_wallet(root_path=self._foxy_root, config=self._config) as wallet_rpc:
             # Select wallet to sync
             await get_wallet(self._foxy_root, wallet_rpc, fingerprint=None)
 
@@ -80,20 +48,9 @@ class PoolJoiner:
             await await_launcher_pool_join_completion(self._foxy_root, joined_launcher_ids)
             print("✅ Pool join completed")
             self._update_foxy_config_plot_nfts_if_required()
-        finally:
-            if wallet_rpc is not None:
-                wallet_rpc.close()
-                await wallet_rpc.await_closed()
-            await self._chia_environment.stop_services(["wallet"])
-            await self._chia_environment.stop_daemon()
 
     def _update_foxy_config_plot_nfts_if_required(self):
-        config = load_config(self._foxy_root, "config.yaml")
-        pool_list: Optional[List[Dict[str, Any]]] = config["pool"].get("pool_list")
-        if pool_list is None:
-            return
         foxy_config = self._foxy_config_manager.load_config()
-        if pool_list == foxy_config.get("plot_nfts"):
-            return
-        foxy_config["plot_nfts"] = pool_list
-        self._foxy_config_manager.save_config(foxy_config)
+        did_update = update_foxy_config_plot_nfts_if_required(root_path=self._foxy_root, foxy_config=foxy_config)
+        if did_update:
+            self._foxy_config_manager.save_config(foxy_config)